requirements.txt
azure-functions==1.20.0
azure-identity==1.17.1
azure-storage-blob==12.19.1
requests==2.32.3
pypdf==4.3.1
python-docx==1.1.2
python-pptx==0.6.23
pandas==2.2.2
pyarrow==16.1.0
faiss-cpu==1.8.0.post1
numpy==1.26.4
